separated votes model from images - copied voting logic from image service to vote service - changed voting endpoint
