+++ app.js
@@ -6,6 +6,7 @@ const winston = require('winston');
 const expressWinston = require('express-winston');
 const postRouter = require('./posts/post.routes');
 const imageRouter = require('./images/image.routes');
+const voteRouter = require('./votes/vote.routes');
 const { protector } = require('./auth/auth.controller');
 const errorHandler = require('./middleware/errorhandler');
 const logger = require('./util/logger');
@@ -34,7 +35,7 @@ app.use(protector);
 app.get('/protected', (req, res) => {
   res.json({ protected: true });
 });
-
+app.use('/votes', voteRouter);
 app.use('/posts', postRouter);
 app.use('/images', imageRouter);
