+++ images/image.controller.js
@@ -4,4 +4,3 @@ const { imageService } = require('./image.service');
 
 exports.uploadImage = imageService.upload();
 exports.getImage = imageService.get();
-exports.upvoteImage = imageService.upvote();
