+++ images/image.routes.js
@@ -4,6 +4,4 @@ const router = Router();
 
 router.route('/:imageId').get(imageContorller.getImage);
 
-router.route('/:imageId/votes').put(imageContorller.upvoteImage);
-
 module.exports = router;
