+++ votes/vote.controller.js
@@ -1,0 +1,3 @@
+const { voteService } = require('./votes.service');
+
+exports.upvote = voteService.upvote();
